refactor(application): import kernel through barrel
